requirements.txt
django==5.2.1
djangorestframework
pyyaml
requests
django-cors-headers